requirements.txt
discord.py 
payos 
aiohttp